+++ CHANGELOG.md
@@ -16,10 +16,13 @@ All user visible changes to `cucumber` crate will be documented in this file. Th
 - Ability for step functions to return `Result`. ([#151])
 - Arbitrary output for `writer::Basic` ([#147])
 - `writer::JUnit` ([JUnit XML report][0110-1]) behind the `output-junit` feature flag ([#147])
+- `writer::Json` ([Json schema][0110-2]) behind the `output-json` feature flag ([#159])
 
 [#147]: /../../pull/147
 [#151]: /../../pull/151
+[#159]: /../../pull/159
 [0110-1]: https://llg.cubic.org/docs/junit
+[0110-2]: https://github.com/cucumber/cucumber-json-schema